[kernel arch x86_64 threading]: move some functions outside of arch.
